Move roll dice logic into roll dice logical and add game rules to roll dice partial view
