+++ WarhammerGameManager.Frontend/Controllers/RollDiceController.cs
@@ -9,19 +9,19 @@ namespace WarhammerGameManager.Frontend.Controllers
 {
     public class RollDiceController : Controller
     {
-        private readonly IGameManagerLogic _gml;
+        private readonly IRollDiceLogic _rdl;
         private readonly IHubContext<GameUpdateHub> _hubContext;
 
-        public RollDiceController(IGameManagerLogic gml, IHubContext<GameUpdateHub> hub)
+        public RollDiceController(IRollDiceLogic rdl, IHubContext<GameUpdateHub> hub)
         {
-            _gml = gml;
+            _rdl = rdl;
             _hubContext = hub;
         }
 
         [HttpGet]
         public async Task<IActionResult> GenerateRollDiceForm()
         {
-            return PartialView("~/Views/RollDice/Partials/RollDiceFormsPartial.cshtml", await _gml.GetRollTypes());
+            return PartialView("~/Views/RollDice/Partials/RollDiceFormsPartial.cshtml", await _rdl.GetRollDiceViewModel());
         }
 
         [HttpPost]
@@ -32,11 +32,11 @@ namespace WarhammerGameManager.Frontend.Controllers
             //If there is no provided GameId, then this is a quick roll
             if (GameId == null)
             {
-                data = await _gml.QuickRollDice(request);
+                data = await _rdl.QuickRollDice(request);
             }
             else
             {
-                data = await _gml.GameRoll(request, GameId.Value);
+                data = await _rdl.GameRoll(request, GameId.Value);
                 await _hubContext.Clients.All.SendAsync("ReceiveRollsUpdates", GameId.Value);
             }
 
@@ -56,11 +56,11 @@ namespace WarhammerGameManager.Frontend.Controllers
             //If there is no provided GameId, then this is a quick roll
             if (GameId == null)
             {
-                response = await _gml.QuickRollDice(request);
+                response = await _rdl.QuickRollDice(request);
             }
             else
             {
-                response = await _gml.GameRoll(request, GameId.Value);
+                response = await _rdl.GameRoll(request, GameId.Value);
                 await _hubContext.Clients.All.SendAsync("ReceiveRollsUpdates", GameId.Value);
             }
